requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
httpx>=0.25.0
docker>=7.0.0
PyPDF2>=3.0.0
qdrant-client>=1.9.0
numpy>=1.21.0
requests>=2.31.0
reportlab>=4.0.0
aiohttp>=3.8.0
PyGithub>=2.0.0
fastmcp>=0.1.0